S15.f.5: extract plugin.ts heartbeat + offload (6488 → 6104 LOC)
Fifth sub-slice of S15.f, bundling two functional clusters in one PR
(per OSS-prep no-micro-PR mandate from user 2026-04-29).

cli/src/core/agt-heartbeat.ts (new, ~156 LOC):
  recordMeshSession(identity, client, target, sessionId, intent, outcome, startedAt)
  agtReconnect(client, isConnected, sandboxName, setConnected, log)
  notifyInboxToMemory(inbox, log)

cli/src/core/agt-offload.ts (new, ~339 LOC):
  runOffloadTask(opts, deps, log) — in-process task loop + workspace harvest
  startProactiveOffloadIfNeeded(deps, log) — boot announce + bg launch
  OffloadDeps interface (meshClient, identity, sandboxName, isConnected,
    offloadInFlight, meshSend, processTaskWithTools)

plugin.ts keeps thin wrappers that capture singleton state via closure
(_offloadDeps()). The ~310 read/write sites elsewhere in plugin.ts
that touch agtMeshClient/agtIdentity/agtSandboxName/etc. are
unchanged — same dep-injection pattern as S15.f.3 mesh transport.

Cumulative S15.f delta:
  pre:  7139
  f.1:  6974  (-165)
  f.2:  6890   (-84)
  f.3:  6648  (-242)
  f.4:  6488  (-160)
  f.5:  6104  (-384)
  Cumulative: -1035. 3104 LOC remaining to §4.2 cap of 3000.

Verification:
  - tsc --noEmit clean / lint 29 (unchanged) / build clean
  - npm test: 454 pass / 2 skipped (same as baseline)

Audit: docs/security-audits/2026-04-29-phase2-hotspot-plugin-cli-f5.md
